+++ CHANGELOG.md
@@ -14,6 +14,7 @@ incremented for features.
 ### Features
 
 * cli: Add keys `members` / `exclude` in config `programs` section ([#546](https://github.com/project-serum/anchor/pull/546)).
+* cli: Allow program address configuration for test command through `clusters.livenet` ([#554](https://github.com/project-serum/anchor/pull/554)).
 
 ### Breaking Changes